Fix "Nothing pasted, nothing gained" in ensime-inf-eval-buffer/definition
The problem is in ensime-inf-eval-region, both ensime-inf-eval-buffer and
ensime-inf-eval-definition try to reuse ensime-inf-eval-region, but
ensime-inf-eval-region uses (use-region-p) to conditionally bypass
the start and end arguments when no region is selected. To fix it,
refactor out the part of the code that deals with sending the string
to the REPL buffer.